+++ docs/src/components/global.jsx
@@ -1,10 +1,11 @@
-import Advanced from "/src/components/advancedMarkdown.astro"
-import Charts from "/src/components/charts.astro"
+import Advanced from "/src/components/advancedMarkdown.astro";
+import Charts from "/src/components/charts.astro";
 import Details from "/src/components/details.astro";
-import GithubActions from "/src/components/github-actions.astro"
-import Modal from '/src/components/modal.astro';
-import BuildInfo from "/src/components/buildinfo.astro"
+import GithubActions from "/src/components/github-actions.astro";
+import Modal from "/src/components/modal.astro";
+import BuildInfo from "/src/components/buildinfo.astro";
+import PatchInfo from "/src/components/patchinfo.astro";
 
-import { Steps, Tabs, TabItem, Card, CardGrid, LinkCard, Aside, Icon } from '@astrojs/starlight/components';
+import { Steps, Tabs, TabItem, Card, CardGrid, LinkCard, Aside, Icon, FileTree } from "@astrojs/starlight/components";
 
-export { Advanced, BuildInfo, Charts, Details, GithubActions, Modal, Steps, Tabs, TabItem, Card, CardGrid, LinkCard, Aside, Icon };
+export { Advanced, BuildInfo, Charts, Details, GithubActions, Modal, Steps, Tabs, TabItem, Card, CardGrid, LinkCard, Aside, Icon, PatchInfo, FileTree };